+++ elasticdl/python/tests/odps_io_test.py
@@ -11,6 +11,7 @@ from elasticdl.python.common.odps_io import ODPSReader, ODPSWriter
 from elasticdl.python.common.odps_recordio_conversion_utils import (
     write_recordio_shards_from_iterator,
 )
+from elasticdl.python.tests.odps_test_utils import create_iris_odps_table
 
 
 @unittest.skipIf(
@@ -34,7 +35,9 @@ class ODPSIOTest(unittest.TestCase):
         self._odps_client = ODPS(
             self._access_id, self._access_key, self._project, self._endpoint
         )
-        self.create_iris_odps_table()
+        create_iris_odps_table(
+            self._odps_client, self._project, self._test_read_table
+        )
 
     def test_read_to_iterator(self):
         reader = ODPSReader(
@@ -117,135 +120,6 @@ class ODPSIOTest(unittest.TestCase):
         self.assertEqual(table.schema.names, columns)
         self.assertEqual(table.schema.types, column_types)
         self.assertEqual(table.to_df().count(), 2)
-
-    def create_iris_odps_table(self):
-        sql_tmpl = """
-        DROP TABLE IF EXISTS {PROJECT_NAME}.{TABLE_NAME};
-        CREATE TABLE {PROJECT_NAME}.{TABLE_NAME} (
-               sepal_length DOUBLE,
-               sepal_width  DOUBLE,
-               petal_length DOUBLE,
-               petal_width  DOUBLE,
-               class BIGINT);
-
-        INSERT INTO {PROJECT_NAME}.{TABLE_NAME} VALUES
-        (6.4,2.8,5.6,2.2,2),
-        (5.0,2.3,3.3,1.0,1),
-        (4.9,2.5,4.5,1.7,2),
-        (4.9,3.1,1.5,0.1,0),
-        (5.7,3.8,1.7,0.3,0),
-        (4.4,3.2,1.3,0.2,0),
-        (5.4,3.4,1.5,0.4,0),
-        (6.9,3.1,5.1,2.3,2),
-        (6.7,3.1,4.4,1.4,1),
-        (5.1,3.7,1.5,0.4,0),
-        (5.2,2.7,3.9,1.4,1),
-        (6.9,3.1,4.9,1.5,1),
-        (5.8,4.0,1.2,0.2,0),
-        (5.4,3.9,1.7,0.4,0),
-        (7.7,3.8,6.7,2.2,2),
-        (6.3,3.3,4.7,1.6,1),
-        (6.8,3.2,5.9,2.3,2),
-        (7.6,3.0,6.6,2.1,2),
-        (6.4,3.2,5.3,2.3,2),
-        (5.7,4.4,1.5,0.4,0),
-        (6.7,3.3,5.7,2.1,2),
-        (6.4,2.8,5.6,2.1,2),
-        (5.4,3.9,1.3,0.4,0),
-        (6.1,2.6,5.6,1.4,2),
-        (7.2,3.0,5.8,1.6,2),
-        (5.2,3.5,1.5,0.2,0),
-        (5.8,2.6,4.0,1.2,1),
-        (5.9,3.0,5.1,1.8,2),
-        (5.4,3.0,4.5,1.5,1),
-        (6.7,3.0,5.0,1.7,1),
-        (6.3,2.3,4.4,1.3,1),
-        (5.1,2.5,3.0,1.1,1),
-        (6.4,3.2,4.5,1.5,1),
-        (6.8,3.0,5.5,2.1,2),
-        (6.2,2.8,4.8,1.8,2),
-        (6.9,3.2,5.7,2.3,2),
-        (6.5,3.2,5.1,2.0,2),
-        (5.8,2.8,5.1,2.4,2),
-        (5.1,3.8,1.5,0.3,0),
-        (4.8,3.0,1.4,0.3,0),
-        (7.9,3.8,6.4,2.0,2),
-        (5.8,2.7,5.1,1.9,2),
-        (6.7,3.0,5.2,2.3,2),
-        (5.1,3.8,1.9,0.4,0),
-        (4.7,3.2,1.6,0.2,0),
-        (6.0,2.2,5.0,1.5,2),
-        (4.8,3.4,1.6,0.2,0),
-        (7.7,2.6,6.9,2.3,2),
-        (4.6,3.6,1.0,0.2,0),
-        (7.2,3.2,6.0,1.8,2),
-        (5.0,3.3,1.4,0.2,0),
-        (6.6,3.0,4.4,1.4,1),
-        (6.1,2.8,4.0,1.3,1),
-        (5.0,3.2,1.2,0.2,0),
-        (7.0,3.2,4.7,1.4,1),
-        (6.0,3.0,4.8,1.8,2),
-        (7.4,2.8,6.1,1.9,2),
-        (5.8,2.7,5.1,1.9,2),
-        (6.2,3.4,5.4,2.3,2),
-        (5.0,2.0,3.5,1.0,1),
-        (5.6,2.5,3.9,1.1,1),
-        (6.7,3.1,5.6,2.4,2),
-        (6.3,2.5,5.0,1.9,2),
-        (6.4,3.1,5.5,1.8,2),
-        (6.2,2.2,4.5,1.5,1),
-        (7.3,2.9,6.3,1.8,2),
-        (4.4,3.0,1.3,0.2,0),
-        (7.2,3.6,6.1,2.5,2),
-        (6.5,3.0,5.5,1.8,2),
-        (5.0,3.4,1.5,0.2,0),
-        (4.7,3.2,1.3,0.2,0),
-        (6.6,2.9,4.6,1.3,1),
-        (5.5,3.5,1.3,0.2,0),
-        (7.7,3.0,6.1,2.3,2),
-        (6.1,3.0,4.9,1.8,2),
-        (4.9,3.1,1.5,0.1,0),
-        (5.5,2.4,3.8,1.1,1),
-        (5.7,2.9,4.2,1.3,1),
-        (6.0,2.9,4.5,1.5,1),
-        (6.4,2.7,5.3,1.9,2),
-        (5.4,3.7,1.5,0.2,0),
-        (6.1,2.9,4.7,1.4,1),
-        (6.5,2.8,4.6,1.5,1),
-        (5.6,2.7,4.2,1.3,1),
-        (6.3,3.4,5.6,2.4,2),
-        (4.9,3.1,1.5,0.1,0),
-        (6.8,2.8,4.8,1.4,1),
-        (5.7,2.8,4.5,1.3,1),
-        (6.0,2.7,5.1,1.6,1),
-        (5.0,3.5,1.3,0.3,0),
-        (6.5,3.0,5.2,2.0,2),
-        (6.1,2.8,4.7,1.2,1),
-        (5.1,3.5,1.4,0.3,0),
-        (4.6,3.1,1.5,0.2,0),
-        (6.5,3.0,5.8,2.2,2),
-        (4.6,3.4,1.4,0.3,0),
-        (4.6,3.2,1.4,0.2,0),
-        (7.7,2.8,6.7,2.0,2),
-        (5.9,3.2,4.8,1.8,1),
-        (5.1,3.8,1.6,0.2,0),
-        (4.9,3.0,1.4,0.2,0),
-        (4.9,2.4,3.3,1.0,1),
-        (4.5,2.3,1.3,0.3,0),
-        (5.8,2.7,4.1,1.0,1),
-        (5.0,3.4,1.6,0.4,0),
-        (5.2,3.4,1.4,0.2,0),
-        (5.3,3.7,1.5,0.2,0),
-        (5.0,3.6,1.4,0.2,0),
-        (5.6,2.9,3.6,1.3,1),
-        (4.8,3.1,1.6,0.2,0);
-        """
-        self._odps_client.execute_sql(
-            sql_tmpl.format(
-                PROJECT_NAME=self._project, TABLE_NAME=self._test_read_table
-            ),
-            hints={"odps.sql.submit.mode": "script"},
-        )
 
     def tearDown(self):
         self._odps_client.delete_table(
